offer state

+++ src/main/java/com/dietiestates2025/dieti/repositories/ImageRepository.java
@@ -8,7 +8,9 @@ import org.springframework.data.repository.query.Param;
 
 import com.dietiestates2025.dieti.model.Image;
 
-public interface ImageRepository extends JpaRepository<Image, Long> {
+public interface ImageRepository extends JpaRepository<Image, Integer> {
     @Query("SELECT i.image FROM Image i JOIN i.property p WHERE p.idProperty = :idProperty")
     List<Image> findByPropertyId(@Param("idProperty") Long idProperty);
+
+
 }
